change the /user/dashboard/edit/:id route to /api/post/edit/:id
