+++ controllers/api/index.js
@@ -1,5 +1,7 @@
 const apiRouters = require('express').Router();
+const postRouter = require('./post');
 
+apiRouters.use('/post',postRouter);
 
 
 module.exports = apiRouters;
